added user management to admin controller
