+++ templates/admin/user/create.html.twig
@@ -10,7 +10,7 @@
     </div>
 
     <div class="bg-white shadow-md rounded-lg p-6 mb-10">
-        <form action="{{ path('user_create') }}" method="POST">
+        <form action="{{ path('admin_user_create') }}" method="POST">
             <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-6">
                 <div>
                     <label for="name" class="block text-sm font-medium text-gray-700 mb-2">Nom :</label>
