+++ templates/admin/user/index.html.twig
@@ -10,7 +10,7 @@
     </div>
 
     <div class="text-center mb-6">
-        <a href="{{ path('user_create') }}" class="px-2 py-1 border-2 border-blue-600 rounded font-medium hover:bg-blue-600 hover:text-white">
+        <a href="{{ path('admin_user_create') }}" class="px-2 py-1 border-2 border-blue-600 rounded font-medium hover:bg-blue-600 hover:text-white">
             Ajouter un utilisateur
         </a>
     </div>
@@ -36,13 +36,13 @@
                         </p>
                     </div>
                     <div class="p-4 bg-gray-50 flex justify-between items-center gap-2">
-                        <a href="{{ path('user_show', { id: user.id }) }}" class="flex-1 px-4 py-2 text-blue-600 border border-blue-600 rounded-md font-medium hover:bg-blue-600 hover:text-white transition-colors text-center">
+                        <a href="{{ path('admin_user_show', { id: user.id }) }}" class="flex-1 px-4 py-2 text-blue-600 border border-blue-600 rounded-md font-medium hover:bg-blue-600 hover:text-white transition-colors text-center">
                             Voir
                         </a>
-                        <a href="{{ path('user_edit', { id: user.id }) }}" class="flex-1 px-4 py-2 text-yellow-600 border border-yellow-600 rounded-md font-medium hover:bg-yellow-600 hover:text-white transition-colors text-center">
+                        <a href="{{ path('admin_user_edit', { id: user.id }) }}" class="flex-1 px-4 py-2 text-yellow-600 border border-yellow-600 rounded-md font-medium hover:bg-yellow-600 hover:text-white transition-colors text-center">
                             Modifier
                         </a>
-                        <form action="{{ path('user_delete', { id: user.id }) }}" method="POST" class="flex-1" onsubmit="return confirm('Voulez-vous vraiment supprimer cet utilisateur ?');">
+                        <form action="{{ path('admin_user_delete', { id: user.id }) }}" method="POST" class="flex-1" onsubmit="return confirm('Voulez-vous vraiment supprimer cet utilisateur ?');">
                             <input type="hidden" name="_method" value="DELETE">
                             <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ user.id) }}">
                             <button type="submit" class="w-full px-4 py-2 text-red-600 border border-red-600 rounded-md font-medium hover:bg-red-600 hover:text-white transition-colors text-center">
