+++ templates/admin/user/show.html.twig
@@ -39,13 +39,13 @@
     </div>
 
     <div class="mt-6 text-center flex justify-center gap-4">
-        <a href="{{ path('user_index') }}" class="px-4 py-2 text-blue-600 border border-blue-600 rounded-md font-medium hover:bg-blue-600 hover:text-white transition-colors">
+        <a href="{{ path('admin_user_index') }}" class="px-4 py-2 text-blue-600 border border-blue-600 rounded-md font-medium hover:bg-blue-600 hover:text-white transition-colors">
             Retour à la liste
         </a>
-        <a href="{{ path('user_edit', { id: user.id }) }}" class="px-4 py-2 text-yellow-600 border border-yellow-600 rounded-md font-medium hover:bg-yellow-600 hover:text-white transition-colors">
+        <a href="{{ path('admin_user_edit', { id: user.id }) }}" class="px-4 py-2 text-yellow-600 border border-yellow-600 rounded-md font-medium hover:bg-yellow-600 hover:text-white transition-colors">
             Modifier
         </a>
-        <form action="{{ path('user_delete', { id: user.id }) }}" method="POST" onsubmit="return confirm('Voulez-vous vraiment supprimer cet utilisateur ?');">
+        <form action="{{ path('admin_user_delete', { id: user.id }) }}" method="POST" onsubmit="return confirm('Voulez-vous vraiment supprimer cet utilisateur ?');">
             <input type="hidden" name="_method" value="DELETE">
             <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ user.id) }}">
             <button type="submit" class="px-4 py-2 text-red-600 border border-red-600 rounded-md font-medium hover:bg-red-600 hover:text-white transition-colors">
