+++ templates/profile/profile.html.twig
@@ -38,7 +38,7 @@
     </div>
 
     <div class="mt-6 text-center">
-        <a href="{{ path('user_profile_edit') }}" class="px-4 py-2 bg-blue-600 text-white rounded-md shadow-md hover:bg-blue-700">
+        <a href="{{ path('profile_edit') }}" class="px-4 py-2 bg-blue-600 text-white rounded-md shadow-md hover:bg-blue-700">
     Modifier mes informations
 </a>
